chore: move auth logic
